Add repro for signal returncode
https://docs.python.org/3.13/library/asyncio-subprocess.html#asyncio.subprocess.Process.returncode
states the following about `returncode`:

> A negative value -N indicates that the child was terminated by signal N (POSIX only).

On some systems, the returncode is -2 as expected. On others, it is 130,
or more generally 128 + signal value. While the docs technically don't
state that signal termination will necessarily result in a negative
returncode, it can easily happen that they are understood that way.
Such behavioral differences are very surprising, and if they are
intentional the docs should warn about them. If they are unintentional,
the should be fixed.